requirements.txt
absl-py==0.12.0
aio-pika==6.8.0
aiofiles==0.6.0
aiohttp==3.7.4
aiormq==3.3.1
APScheduler==3.7.0
asgiref==3.3.4
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
bidict==0.21.2
blis==0.7.4
boto3==1.17.68
botocore==1.20.68
cachetools==4.2.2
catalogue==2.0.4
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorclass==2.2.0
coloredlogs==15.0
colorhash==1.0.3
contextvars==2.4
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.5
dataclasses==0.8
decorator==4.4.2
Django==3.2.2
djangorestframework==3.12.4
dm-tree==0.1.6
dnspython==1.16.0
docopt==0.6.2
es-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.0.0/es_core_news_lg-3.0.0-py3-none-any.whl
fbmessenger==6.0.0
future==0.18.2
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.37.1
h11==0.9.0
h5py==2.10.0
httpcore==0.11.1
httplib2==0.19.1
httptools==0.2.0
httpx==0.15.4
humanfriendly==9.1
idna==2.10
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==4.0.1
Jinja2==2.11.3
jmespath==0.10.0
joblib==0.15.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==2.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mattermostwrapper==2.2
multidict==4.7.6
murmurhash==1.0.5
networkx==2.5.1
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
pathy==0.5.2
Pillow==5.4.1
preshed==3.0.5
prompt-toolkit==2.0.10
protobuf==3.15.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
pydot==1.4.2
PyJWT==2.1.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
pyTelegramBotAPI==3.7.7
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
questionary==1.9.0
rasa==2.6.0
rasa-sdk==2.6.0
redis==3.5.3
regex==2020.9.27
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rocketchat-API==1.15.0
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
sanic==20.9.0
Sanic-Cors==0.10.0.post3
sanic-jwt==1.5.0
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.24.2
scipy==1.5.4
sentry-sdk==0.19.5
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smart-open==3.0.0
sniffio==1.2.0
spacy==3.0.6
spacy-legacy==3.0.5
SQLAlchemy==1.3.24
sqlparse==0.4.1
srsly==2.4.1
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.2
tensorflow-addons==0.12.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-probability==0.11.1
tensorflow-text==2.3.0
termcolor==1.1.0
terminaltables==3.1.0
thinc==8.0.3
threadpoolctl==2.1.0
tqdm==4.59.0
twilio==6.50.1
typeguard==2.12.0
typer==0.3.2
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
Unidecode==1.2.0
urllib3==1.26.4
uvloop==0.14.0
wasabi==0.8.2
wcwidth==0.2.5
webexteamssdk==1.6
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
